Rename the index app to be the dashboard
"Dashboard" is a more descriptive name for what this app
will be doing so this change renames the index app to be
the new dashboard app.
