+++ app/dashboard/controllers.js
@@ -1,0 +1,8 @@
+module.exports = {
+  get: (req, res) => {
+    const params = {
+      pageTitle: 'Upcoming moves',
+    }
+    res.render('dashboard/dashboard', params)
+  },
+}
