+++ app/dashboard/dashboard.njk
@@ -1,14 +1,14 @@
 {% extends "layouts/govuk.njk" %}
 
 {% block pageTitle %}
-  {{ message }}
+  {{ pageTitle }}
 {% endblock %}
 
 {% block content %}
 <div class="govuk-grid-row">
   <div class="govuk-grid-column-two-thirds">
     <h1 class="govuk-heading-xl">
-      {{ message }}
+      {{ pageTitle }}
     </h1>
   </div>
 </div>
