--- app/index/controllers.js
@@ -1,8 +1,0 @@
-module.exports = {
-  get: (req, res) => {
-    const params = {
-      message: `Hello, world!`,
-    }
-    res.render('index/index', params)
-  },
-}
